Add support for .jxr saved from Windows Game Bar
The Windows Game Bar saves HDR screenshots in JPEG XR as well,
however it uses a different pixel format that wasn't previously
working.

These files are using 16-bit half float for each channel, but
are also RGBA so it was pretty straightforward to add an adapter,
using the 'half' crate to support the f16 <-> f32 conversions.

Settings that match the SDR variant for the sample file on issue #1
reasonably closely (but not exactly) are:

  --exposure=-2 --hdr-max=480 --saturation=32768

Fixes #1